user

+++ src/main/java/com/ms/ware/online/solution/service/BranchServiceImpl.java
@@ -3,6 +3,8 @@ package com.ms.ware.online.solution.service;
 
 import com.ms.ware.online.solution.config.exceptiion.CustomException;
 import com.ms.ware.online.solution.config.exceptiion.PermissionDeniedException;
+import com.ms.ware.online.solution.config.security.AuthenticatedUser;
+import com.ms.ware.online.solution.config.security.AuthenticationFacade;
 import com.ms.ware.online.solution.entity.Branch;
 import com.ms.ware.online.solution.repository.BranchRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,13 +17,17 @@ public class BranchServiceImpl implements BranchService {
 
     @Autowired
     private BranchRepository repository;
+    @Autowired
+    private AuthenticationFacade facade;
 
     @Override
-    public void save(Branch user) {
-        if (user.getId() == null) {
-            user.setId(repository.findNextId());
+    public void save(Branch obj) {
+        AuthenticatedUser user = facade.getAuthentication();
+        if (user.getBranch() != 1 || !user.getUsertype().equalsIgnoreCase("ADM")) throw new PermissionDeniedException();
+        if (obj.getId() == null) {
+            obj.setId(repository.findNextId());
         }
-        repository.save(user);
+        repository.save(obj);
     }
 
     @Override
@@ -36,6 +42,8 @@ public class BranchServiceImpl implements BranchService {
 
     @Override
     public void deleteById(Integer id) {
+        AuthenticatedUser user = facade.getAuthentication();
+        if (user.getBranch() != 1 || !user.getUsertype().equalsIgnoreCase("ADM")) throw new PermissionDeniedException();
         if (id == 1) throw new PermissionDeniedException();
         repository.deleteById(id);
     }
